Optimize Mihon workflow state handling
